[iOS] SPref example & refactor folder
