+++ iosApp/iosApp/features/explore/view/ExploreView.swift
@@ -1,0 +1,16 @@
+//
+//  ExploreView.swift
+//  iosApp
+//
+//  Created by Huu Tran on 26/04/2023.
+//  Copyright © 2023 orgName. All rights reserved.
+//
+
+import Foundation
+import SwiftUI
+
+struct ExploreView: View {
+    var body: some View {
+        Text("Explore")
+    }
+}
