+++ iosApp/iosApp/features/portfolio/view/PortfolioView.swift
@@ -1,0 +1,15 @@
+//
+//  PortfolioView.swift
+//  iosApp
+//
+//  Created by Huu Tran on 26/04/2023.
+//  Copyright © 2023 orgName. All rights reserved.
+//
+
+import SwiftUI
+
+struct PortfolioView: View {
+    var body: some View {
+        Text("Portfolio")
+    }
+}
