+++ iosApp/iosApp/features/search/view/SearchView.swift
@@ -1,0 +1,16 @@
+//
+//  SearchView.swift
+//  iosApp
+//
+//  Created by Huu Tran on 26/04/2023.
+//  Copyright © 2023 orgName. All rights reserved.
+//
+
+import Foundation
+import SwiftUI
+
+struct SearchView: View {
+    var body: some View {
+        Text("Search")
+    }
+}
